@@ -1,16 +1,16 @@
 package com.infiauto.datastr;
 
-import java.io.Serializable;
 import java.util.BitSet;
 
 /**
+ * Extension of the JDK BitSet class that provides for specifying a fixed length
+ * along with the ability to de/serialized.
  * @author Infinite Automata
  */
 public class FixedLengthBitSet
-extends BitSet
-implements Serializable {
-	private static final long serialVersionUID = 3567697288265161400L;
-	private int fixed_length;
+        extends BitSet {
+
+    private int fixed_length;
 
     public FixedLengthBitSet(int length) {
         this.fixed_length = length;
@@ -19,51 +19,39 @@ implements Serializable {
     public int fixedLength() {
         return fixed_length;
     }
-
-    @Override
-    public boolean equals(Object obj) {
-/*        if(!(obj instanceof FixedLengthBitSet)) {
-            return false;
-        }*/
-        FixedLengthBitSet a = (FixedLengthBitSet) obj;
-        return super.equals(a) && (this.fixed_length == a.fixed_length);
-/*        if ((compareTo (a) == 0) != super.equals(obj) ) {
-        		System.out.println (this.toString() + "/" + super.toString() + "/" + a.toString());
-        		
-        		System.out.println (this.fixedLength());
-        		System.out.println (a.fixedLength());
-        		
-        		throw new RuntimeException("FU!");
-       		
+    
+    public FixedLengthBitSet substring(int start, int length) {
+        assert(start + length <= fixed_length);
+        
+        FixedLengthBitSet result = new FixedLengthBitSet(length);
+        for(int i = 0; i < length; i++) {
+            result.set(i, this.get(start + i));
         }
-        return compareTo(a) == 0;*/
+        return result;
     }
     
-    
-    
-    /*@Override
-    public int hashCode() {
-        return toString().hashCode();
-    }*/
-
     @Override
-	public void set(int bitIndex) {
-    	if (bitIndex >= fixed_length)
-    		throw new RuntimeException ("Низя!");   		
-
-		super.set(bitIndex);
-	}
-
-/*	@Override
-    public int compareTo(FixedLengthBitSet bits) {
-        return toString().compareTo(bits.toString());
-    }*/
+    public boolean equals(Object obj) {
+        if (!(obj instanceof FixedLengthBitSet)) {
+            return false;
+        }
+        FixedLengthBitSet a = (FixedLengthBitSet) obj;
+        if(fixedLength() != a.fixedLength()) {
+            return false;
+        }
+        return super.equals(obj);
+    }
+    
+    @Override
+    public int hashCode() {
+        return super.hashCode() ^ fixed_length * 4703;
+    }
 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder(fixed_length);
-        
-        for(int i = (fixed_length - 1); i >= 0; i--) {
+
+        for (int i = (fixed_length - 1); i >= 0; i--) {
             sb.append(get(i) ? '1' : '0');
         }
 
